Move taxonomy API classes from wastd/api.py into taxonomy application.
